requirements.txt
torch
transformers
peft
faiss-cpu
flask
flask-cors
onnx
onnxruntime
scikit-learn
sentence-transformers
rank-bm25
numpy
pandas
